% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{classify.fcs}
\alias{classify.fcs}
\title{Classify particles from an FCS dataframe.}
\usage{
classify.fcs(fcs, gates.log)
}
\arguments{
\item{fcs}{FCS data frame.}

\item{gates.log}{A gating scheme from the function "add.manual.classification()" or "add.auto.classification()"}
}
\value{
List of per particle classifications
}
\description{
Classify particles from an FCS dataframe using a gating scheme provided by gates.log.
}
\examples{
\dontrun{
opp <- classify.fcs(fcs, gates.log)
}
}
