% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{cmap_convert}
\alias{cmap_convert}
\title{Format Influx data into CAMP compatible format.}
\usage{
cmap_convert(data, cruise, project, version = "v1.0")
}
\arguments{
\item{data}{data from Influx.}

\item{cruise}{cruise name, if any (otherwise NA).}

\item{project}{cruise name, if any (otherwise NA).}

\item{version}{Version of the dataset.}
}
\value{
None
}
\description{
Format Influx data into CAMP compatible format.
}
\examples{
\dontrun{
csv_convert(db, meta, path)
}
}
