% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{manual.classify}
\alias{manual.classify}
\title{Classify particles based on manually defined population gates.}
\usage{
manual.classify(fcs, params, popname)
}
\arguments{
\item{fcs}{FCS data frame.}

\item{params}{Named list of gating parameters. Must contain a params$poly
entry with polygon definitions.}

\item{popname}{Name of the population}
}
\value{
List of per particle classifications.
}
\description{
Classify particles based on manually defined population gates.
}
\examples{
\dontrun{
vct <- manual.classify(opp, gates.log, "beads")
}
}
