% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{plot_vct_cytogram}
\alias{plot_vct_cytogram}
\title{Plot cytogram with particles colored by population.}
\usage{
plot.vct.cytogram(fcs, para.x = "scatter", para.y = "red")
}
\arguments{
\item{fcs}{FCS data frame.}

\item{para.x}{Channel to use as x axis.}

\item{para.y}{Channel to use as y axis.}

\item{...}{Additional parameters for plot()}
}
\value{
None
}
\description{
Plot cytogram with particles colored by population.
}
