% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{read_influx}
\alias{read_influx}
\title{read a fcs file}
\usage{
fcs <- read.influx(fcs_file)
}
\arguments{
\item{fcs_file}{path to the FCS file.}

\item{tranformation}{whether or not to transform the data. Default is TRUE}

\item{...}{Additional parameters for flowCore::read.FCS()}
}
\value{
a tibble dataframe
}
\description{
read a fcs file
}
