% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{xls_convert}
\alias{xls_convert}
\title{Format Influx data into an Excel spreadsheet, along with metadata.}
\usage{
xls_convert(data, project, cruise, version = "v1.0")
}
\arguments{
\item{data}{data from Influx.}

\item{project}{Version of the dataset.}

\item{cruise}{cruise name, if any (otherwise NA).}

\item{version}{Version of the dataset.}
}
\value{
None
}
\description{
Format Influx data into an Excel spreadsheet, along with metadata.
}
\examples{
\dontrun{
csv_convert(db, meta, path)
}
}
